% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se_functions.R
\name{bootstrap_mean}
\alias{bootstrap_mean}
\title{Bootstrap mean and difference in means}
\usage{
bootstrap_mean(data = data, i, object, contrast1, contrast2 = NULL)
}
\arguments{
\item{data}{the original data frame}

\item{i}{indices}

\item{object}{An object of class "polr"}

\item{contrast1}{First contrast vector for calculating a mean based on
covariate values}

\item{contrast2}{Second contrast vector for calculating a mean based on
covariate values}
}
\value{
A mean associated with contrast 1 or two means and their difference
based on constrasts 1 and 2
}
\description{
This function is called when bootstrapping means and differences in means
}
\keyword{internal}
