% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calculate_cumprobs}
\alias{calculate_cumprobs}
\title{Calculate cumulative probabilities}
\usage{
calculate_cumprobs(contrast, n_int, theta)
}
\arguments{
\item{contrast}{Contrast vector for calculating a mean based on
covariate values}

\item{n_int}{Number of intercept parameters}

\item{theta}{Vector of intercept parameters and regression coefficients}
}
\value{
The cumulative probabilities associated with the specified contrast
}
\description{
Calculate cumulative probabilities
}
\keyword{internal}
