% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se_functions.R
\name{calculate_delta_method_se}
\alias{calculate_delta_method_se}
\title{Calculate Delta Method Standard Error}
\usage{
calculate_delta_method_se(object, contrast, n_int, theta, B)
}
\arguments{
\item{object}{An object of class "polr"}

\item{contrast}{Contrast vector for calculating a mean based on
covariate values}

\item{n_int}{Number of intercept parameters}

\item{theta}{Vector of intercepts and regression coefficients}

\item{B}{Matrix for converting cumulative probabilities to probabilities}
}
\value{
The standard error of mean associated with the specified contrast
vector
}
\description{
Calculate Delta Method Standard Error
}
\keyword{internal}
