% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se_functions.R
\name{calculate_difference_and_se}
\alias{calculate_difference_and_se}
\title{Calculate the difference of two means and their standard error}
\usage{
calculate_difference_and_se(
  object,
  contrast1,
  contrast2,
  n_int,
  theta,
  B,
  yvals
)
}
\arguments{
\item{object}{An object of class "polr"}

\item{contrast1}{First contrast vector for calculating a mean based on
covariate values}

\item{contrast2}{Second contrast vector for calculating a mean based on
covariate values}

\item{n_int}{Number of intercept parameters}

\item{theta}{A vector of intercept parameters and regression coefficients}

\item{B}{Matrix for converting cumulative probabilities to probabilities}
}
\value{
A vector consisting of the difference in means, its SE, Z-statistic,
and p-value
}
\description{
Calculate the difference of two means and their standard error
}
\keyword{internal}
