% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se_functions.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Calculate mean}
\usage{
calculate_mean(object, contrast, n_int, theta, B)
}
\arguments{
\item{object}{An object of class "polr"}

\item{contrast}{Contrast vector for calculating a mean based on
covariate values}

\item{n_int}{Number of intercept parameters}

\item{theta}{Vector of intercept parameters and regression coefficients}

\item{B}{Matrix for converting cumulative probabilities to probabilities}
}
\value{
The mean associated with the specified contrast
}
\description{
Calculate mean
}
\keyword{internal}
