% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calculate_normal_ci}
\alias{calculate_normal_ci}
\title{Calculate a normal-based confidence interval}
\usage{
calculate_normal_ci(output, level)
}
\arguments{
\item{output}{An object of class "ordinalTools"}

\item{level}{Confidence interval level}
}
\value{
Upper and lower bounds of a confidence interval
}
\description{
Calculate a normal-based confidence interval
}
\keyword{internal}
