% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calculate_z_statistics}
\alias{calculate_z_statistics}
\title{Calculate z-statistics}
\usage{
calculate_z_statistics(mean, se)
}
\arguments{
\item{mean}{Mean value based on multiplying yvalues by their associated probabilities}

\item{se}{Standard Error. Either based on delta method or bootstrapping}
}
\value{
Vector of mean, SE, z statistic, and p-value
}
\description{
Calculate z-statistics
}
\keyword{internal}
