% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{coef.ordinalTools}
\alias{coef.ordinalTools}
\title{Extract model coefficients}
\usage{
\method{coef}{ordinalTools}(object, ...)
}
\arguments{
\item{object}{an object of class "ordinalTools"}
}
\value{
Coefficients extracted from the model object
}
\description{
Extract model coefficients
}
