% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_B_matrix}
\alias{create_B_matrix}
\title{Create B matrix}
\usage{
create_B_matrix(n_int)
}
\arguments{
\item{n_int}{Number of intercept parameters}
}
\value{
A matrix for turning cumulative probabilities into probabilities
}
\description{
Create B matrix
}
\keyword{internal}
