% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{create_pickoff_matrix}
\alias{create_pickoff_matrix}
\title{Create pickoff matrix}
\usage{
create_pickoff_matrix(contrast, n_int)
}
\arguments{
\item{contrast}{Contrast for specifying mean of interest}

\item{n_int}{Number of intercept parameters}
}
\value{
A matrix that can be used to calculate cumulative probabilities
}
\description{
Create pickoff matrix
}
\keyword{internal}
