% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cumprobs_to_probs}
\alias{cumprobs_to_probs}
\title{Turn cumulative probabilities to probabilities}
\usage{
cumprobs_to_probs(cumprobs, n_int, B)
}
\arguments{
\item{cumprobs}{Vector of cumulative probabilities}

\item{n_int}{Number of intercept parameters}

\item{B}{B matrix}
}
\value{
A vector of probabilities
}
\description{
Turn cumulative probabilities to probabilities
}
\keyword{internal}
