% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format.perc}
\alias{format.perc}
\title{Format a probability as a percentage}
\usage{
\method{format}{perc}(probs, digits)
}
\arguments{
\item{probs}{Probabilty value}

\item{digits}{The number of digits}
}
\value{
Formatted percentages
}
\description{
Format a probability as a percentage
}
\keyword{internal}
