% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{format_perc}
\alias{format_perc}
\title{Format a probability as a percentage}
\usage{
format_perc(probs, digits)
}
\arguments{
\item{probs}{Probability value}

\item{digits}{The number of digits}
}
\value{
Formatted percentage
}
\description{
Format a probability as a percentage
}
\keyword{internal}
