% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_and_se_functions.R
\name{get_mean_and_se}
\alias{get_mean_and_se}
\title{Get mean and standard error from an ordinal model via
the delta method}
\usage{
get_mean_and_se(object, contrast, n_int, theta, B)
}
\arguments{
\item{object}{An object of class ``polr"}

\item{contrast}{Contrast vector for calculating a mean based on
covariate values}

\item{n_int}{Number of intercept parameters}

\item{theta}{Vector of intercept parameters and regression coefficients}
}
\value{
A data frame consisting of a mean, its SE, z-statistic, and p-value
}
\description{
This function is a wrapper function that uses an object of class "polr" and a specified contrast
vector to calculate a mean, its SE, and Z-statistics
}
\keyword{internal}
