% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_gee_mean_function.R
\name{ordinal_gee_means}
\alias{ordinal_gee_means}
\title{Calculate means from an ordinal GEE model}
\usage{
ordinal_gee_means(
  object,
  data = data,
  conf.level = 0.95,
  contrast1 = rep(0, length(coef(object))),
  contrast2 = NULL,
  se.type = "delta",
  R = 500
)
}
\arguments{
\item{object}{An object of class "LORgee"}

\item{data}{A data frame.}

\item{conf.level}{Confidence interval level}

\item{contrast1}{A vector the same length as the number of coefficients in}

\item{contrast2}{A vector the same length as the number of coefficients in}

\item{se.type}{Method for calculating standard errors. Options are
the delta method (se.type="delta") or bootstrapping (se.type="bootstrap")}

\item{R}{The number of bootstrap samples}
}
\value{
An object of class "ordinalTools. This has components
\itemize{
\item call the matched call
\item output Parameter estimates, their standard errors, test statistics, and p-values
\item conf.int Confidence interval
\item se.type Method used to calculate standard errors
}
}
\description{
Calculate means from an ordinal GEE model
}
\details{
This function uses the parameter estimates from an proportional odds
GEE logistic regression model that was fit using the multgee::ordLORgee function to obtain
mean values for a given set of coefficient contrasts. Up to two constrasts
can be specified. Standard errors are calculated either using the delta
method or bootstrapping.
}
\examples{
\dontrun{
fit <- ordinal_means(multgee.fit)
summary(fit)

# Two coefficient constrasts and their difference
fit <- ordinal_means(multgee.fit, contrast1=c(0,0,0), contrast2=c(0,0,0))
summary(fit)

# Two coefficient constrasts and their difference
# Bootstrap standard errors
fit <- ordinal_means(multgee.fit, data=data, contrast1=c(0,0,0),
       contrast2=c(0,0,0), type="bootstrap")
summary(fit)
}
}
