% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalTools-package.R
\docType{package}
\name{ordinalTools-package}
\alias{ordinalTools}
\alias{ordinalTools-package}
\title{ordinalTools: Tools for Working with Output from Ordinal Models}
\description{
This package provides functions for working with output from ordinal regression models.
}
\author{
\strong{Maintainer}: Juned Siddique \email{siddique@northwestern.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
