% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.ordinalTools}
\alias{print.ordinalTools}
\title{Print Coefficient Matrices}
\usage{
\method{print}{ordinalTools}(x, ...)
}
\arguments{
\item{x}{an object of class "ordinalTools"}

\item{...}{other arguments}
}
\value{
a matrix with columns for the estimated coefficient,
its standard error, z-statistic and corresponding (two-sided) p-value.
}
\description{
Print Coefficient Matrices
}
\examples{
\dontrun{
fit <- ordinal_means(polr.fit)
print(fit)
}
}
