% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{reorder_vcov}
\alias{reorder_vcov}
\title{Reorder a variance-covariance matrix}
\usage{
reorder_vcov(old, new_order = 1:nrow(old))
}
\arguments{
\item{old}{The original variance-covariance matrix}

\item{new_order}{The desired order in terms of the original columns.
Default is the original order: new_order=1:nrow(old).}
}
\value{
A reordered variance-covariance matrix
}
\description{
Reorder a variance-covariance matrix
}
\keyword{internal}
