% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smoke14}
\alias{smoke14}
\title{Daily smoking data}
\format{
A data frame with 383 rows and 3 columns:
\describe{
\item{Cig14dRate}{Average cigarettes per day over 14 days}
\item{GenderF}{Indicator variable for female gender}
\item{CigNDSS}{Nicotine Dependence Symptom Scale}

}
}
\source{
< https://doi.org/10.1046/j.1360-0443.95.11s3.7.x>
}
\usage{
data(smoke14)
}
\description{
A dataset containing average number of cigarettes per day over 14
days in a sample of 383 participants. The variables are as follows:
}
\keyword{datasets}
