% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summary.ordinalTools}
\alias{summary.ordinalTools}
\title{Summarize results of an object of class "ordinalTools"}
\usage{
\method{summary}{ordinalTools}(
  object,
  digits = max(3, getOption("digits") - 2),
  signif.stars = TRUE,
  dig.tst = max(1, min(5, digits - 1)),
  ...
)
}
\arguments{
\item{object}{An object of class "ordinalTools" for which a summary is desired}

\item{digits}{minimum number of significant digits to be used for most numbers}

\item{signif.stars}{ogical; if TRUE, P-values are additionally encoded
visually as ‘significance stars’ in order to help scanning of long coefficient tables.}

\item{dig.tst}{minimum number of significant digits for the test statistics}

\item{...}{other arguments}
}
\value{
Parameter estimates, their SEs, test statistic, and p-value
}
\description{
Summarize results of an object of class "ordinalTools"
}
\examples{
\dontrun{
fit <- ordinal_means(polr.fit)
summary(fit)
}
}
