% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinalTools-package.R
\docType{package}
\name{ordinalTools-package}
\alias{ordinalTools}
\alias{ordinalTools-package}
\title{ordinalTools: Tools for Working with Output from Ordinal Models}
\description{
This package provides functions for working with output from ordinal regression models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/junedsiddique/ordinalTools}
  \item Report bugs at \url{https://github.com/junedsiddique/ordinalTools/issues}
}

}
\author{
\strong{Maintainer}: Juned Siddique \email{siddique@northwestern.edu} (\href{https://orcid.org/0000-0002-1501-4152}{ORCID})

}
\keyword{internal}
