% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.ordinalTools}
\alias{print.ordinalTools}
\title{Print Coefficient Matrices}
\usage{
\method{print}{ordinalTools}(object, ...)
}
\arguments{
\item{object}{an object of class "ordinalTools"}
}
\value{
a matrix with columns for the estimated coefficient,
its standard error, z-statistic and corresponding (two-sided) p-value.
}
\description{
Print Coefficient Matrices
}
